Update blockchain.bib
We have withdrawn our paper in 3 Mar 2021. Now we have revised the paper with great changes and want to submit to some journal, and we want to withdraw the old version from Blocks & Chains Bibliography since it may affect the result of plagiarism detection. 

Could you help me delete the cashed pdf? thanks!